fix(tunnel): close Milestone D Phase 1 — feature gate, clippy, and test debt
- Stub WireGuard kernel backend (wireguard_control API incompatible)
- Remove broken tun-rs platform module (missing dep, API mismatches)
- Fix clippy: unnecessary cast in messages.rs, too_many_arguments in runtime.rs
- Fix stats parsers: parse_transfer_bytes decimal+unit, parse_handshake_time continue-on-non-numeric
- Correct test expectations from SI to binary (MiB) units
- Update TUNNELS.md: WireGuard is feature-gated/stubbed, not removed

75/75 tests pass, clippy clean across default/wireguard/all-features profiles.

+++ crates/synvoid-tunnel/src/quic/messages.rs
@@ -281,7 +281,7 @@ impl DatagramMessage {
     }
 
     pub fn encoded_size(&self) -> usize {
-        crate::serialization::serialized_size(self) as usize
+        crate::serialization::serialized_size(self)
     }
 
     pub fn is_fragmented(&self) -> bool {
